Added the abstraction for a M:TG card
    The only really messy thing left in this abstraction is the fact that all
the cards are subclasses of Client. This means that the card needs to be init'd
with the GameInstance, and somehow needs to be aware of where in play it is.

    Normally, this isn't a problem because actions such as exiling and
discarding are public anyway. Drawing and searching however do represent a
problem because the card has to be alerted (a signal is preferred) but the
enemy/AI does NOT need to be told what card was drawn.