Reverted the EssentialMagic scripts, tweaks to gatherer and client
	After much testing and several hours spent revising an improved
version of the essential magic downloading script it became obvious that
my time was better spent elsewhere so I reverted to the known working
script which was used on the first run against EssentialMagic, made some
tweaks to fix a parser bug in the client and about 87 different UnicodeError
issues.

	Some touches to the standard client object as well, altering how
self.upkeep() is called.